Extract shared code from SHM streaming executables into reusable headers
Factor out ~700 lines of duplicated code from m2sdr_rx_stream_shm.c and
m2sdr_tx_stream_shm.c into two new headers:

- m2sdr_shm.h: Unified 64-byte SHM ring buffer layout with atomic
  accessors, shm_create(), shm_open_existing(), and shm_destroy()
- m2sdr_common.h: Signal handling, AD9361 platform stubs, AGC helpers,
  and shared RF init (SI5351, AD9361 power-up, SPI, bitmode, etc.)

The unified SHM header layout ensures RX and TX use identical offsets,
enabling a future full-duplex executable. RX and TX use separate SHM
files so the error_count field (bytes 16-23) safely serves as
overflow_count in RX and underflow_count in TX.
